refactor(context): reorganize context files
