+++ elements/rh-context-provider/demo/rh-context-provider.html
@@ -126,25 +126,3 @@
     <span>Description Placeholder</span>
   </rh-stat>
 </rh-context-provider>
-
-<rh-context-provider color-palette="base">
-  <h2>Base</h2>
-  <p>Here are some components that should be aware of their context:</p>
-
-  <rh-cta>Hello World</rh-cta>
-  <rh-spinner size="sm"></rh-spinner>
-  <rh-blockquote align="center">
-    <p>In open source, we feel strongly that to really do something well, you have to get a lot of people involved.</p>
-    <span slot="author">Linus Torvalds</span>
-    <span slot="title">Software Engineer</span>
-  </rh-blockquote>
-  <rh-stat>
-    <svg slot="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 36 36">
-      <path d="M17.37 8v10a.63.63 0 0 0 1.25 0V8a.63.63 0 0 0-1.25 0Zm7 0v7a.63.63 0 0 0 1.25 0V8a.63.63 0 0 0-1.25 0Zm-14 0v12a.63.63 0 0 0 1.25 0V8a.63.63 0 0 0-1.25 0ZM31 17.89a.63.63 0 0 0-.63.62v11.87H5.62v-2.93a.63.63 0 0 0-1.25 0V31a.63.63 0 0 0 .63.62h26a.62.62 0 0 0 .62-.62V18.51a.62.62 0 0 0-.62-.62Z"/>
-      <path d="M5 21a.63.63 0 0 0 .62-.63V5.62h24.75V9a.63.63 0 0 0 1.25 0V5a.62.62 0 0 0-.62-.62H5a.63.63 0 0 0-.63.62v15.36A.63.63 0 0 0 5 21Zm27.35-9.24a.62.62 0 0 0-.87.17C28.73 16 21.5 22.93 4 23.27a.63.63 0 0 0 0 1.25c18.07-.34 25.64-7.61 28.52-11.9a.62.62 0 0 0-.17-.86Z"/>
-    </svg>
-    <span slot="statistic">Statistic Placeholder</span>
-    <span>Description Placeholder</span>
-  </rh-stat>
-</rh-context-provider>
-
