+++ elements/rh-secondary-nav/rh-secondary-nav-dropdown.ts
@@ -8,7 +8,7 @@ import { bound, observed } from '@patternfly/pfe-core/decorators.js';
 import { SlotController } from '@patternfly/pfe-core/controllers/slot-controller.js';
 import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
 
-import { colorContextProvider } from '../../lib/context/color.js';
+import { colorContextProvider } from '../../lib/context/decorators.js';
 
 import { RhSecondaryNavMenu } from './rh-secondary-nav-menu.js';
