+++ lib/context/types.ts
@@ -1,0 +1,28 @@
+/**
+ * A `ColorPalette` is a collection of specific color values
+ * Choosing a palette sets both color properties and, if the component is a context provider,
+ * implies a color theme for descendents.
+ *
+ * `ColorPalette` is associated with the `color-palette` attribute
+ */
+export type ColorPalette = (
+    | 'base'
+    | 'accent'
+    | 'complement'
+    | 'lighter'
+    | 'lightest'
+    | 'darker'
+    | 'darkest'
+  );
+
+/**
+   * A Color theme is a context-specific restriction on the available color palettes
+   *
+   * `ColorTheme` is associated with the `on` attribute and the `--context` css property
+   */
+export type ColorTheme = (
+    | 'dark'
+    | 'light'
+    | 'saturated'
+  );
+
